pyproject.toml
[project]
name = "law-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agents>=1.4.0",
    "asyncio>=3.4.3",
    "openai>=1.84.0",
    "openai-agents>=0.0.16",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
